refactored user seeds
