+++ .gitignore
@@ -1,3 +1,3 @@
 config.env
 node_modules/
-seeds.js
+user_seeds.js
